code-review: run each pass of the division unconditionally, and say once what serves a read
Findings the whole suite settled.

Each pass of the division ran behind a flag saying whether the pass
before it had changed anything, and five `changed |=` sites in
`keep_each_global_where_it_is_read` kept that flag true. Dropping one
would have left `published`, `published_here` and the pruning reading a
state a later pass had replaced — the same shape as #608. Each pass now
runs, and the last one is what the division reports. LangArena at
`-O max` emits byte-identical code and builds 1.2% slower.

Whether a unit can serve a read of a global was written out in the
division's assertion and again in `implement_rc_global`, which decides
the accessor from it. `RcGlobalInit::serves_reads` says it once. The
assertion reads its readers from `units_reading_each_global` rather
than walking the bodies a second time.

`generated_code_hash` destructures `DividedProgram`, so a field added
to the division is a compile error there until the digest says what it
does to a unit's code.

`collect_mentions` moves to `rc_ir::ast`, beside `for_each_node` and
`for_each_var`, which walk the same tree.
